% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-pred.R
\name{cat_majority}
\alias{cat_majority}
\title{FUNCTION: cat_majority}
\usage{
cat_majority(data, pred, col = "program", predict = "range")
}
\description{
This function takes in data and the desired program to predict mutation
scores for and returns the accuracy when predicting majority occurrence
}

