% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cat_pred_bench_all}
\alias{cat_pred_bench_all}
\title{FUNCTION: cat_pred_bench_all}
\usage{
cat_pred_bench_all(trials)
}
\description{
This function runs experimentation
prediction on all programs using all
machine learning techniques
}

