% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat-pred.R
\name{cat_predict}
\alias{cat_predict}
\title{FUNCTION: cat_predict}
\usage{
cat_predict(data, method, pred)
}
\description{
This function takes in data
predicts the desired value using the
desired machine learning algorithm
}

