% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat-predict.R
\name{cat_prediction}
\alias{cat_prediction}
\title{FUNCTION: cat_prediction}
\usage{
cat_prediction(data, method, pred, kfolds = 10, repeats = 1)
}
\description{
This function is supports basic categorical prediction
}

