% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-pred.R
\name{cat_rand}
\alias{cat_rand}
\title{FUNCTION: cat_rand}
\usage{
cat_rand(data, pred, progcol = "program", predcol = "range")
}
\description{
This function takes in data and the desired program to predict mutation
scores for and returns the accuracy when predicting random occurrences
}

