% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{cat_sample}
\alias{cat_sample}
\title{FUNCTION: cat_sample}
\usage{
cat_sample(data, sampling, target)
}
\arguments{
\item{data}{to manipulate}

\item{sampling}{method either up or down}

\item{target}{variable to manipulate data to}
}
\description{
This function takes in data and samples it to the desired
technique
}

