% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cpred_by_program}
\alias{cpred_by_program}
\title{FUNCTION: cpred_by_program}
\usage{
cpred_by_program(method = "C5.0", pred = "bpmail", data = totalCDCatEvo)
}
\arguments{
\item{method}{Machine learning algorithm}

\item{pred}{Program to predict MS}

\item{data}{All programs with MS}
}
\description{
User prediction when splitting the data frame to
predict for one program and train on all the rest
}

