% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_npreds}
\alias{graph_npreds}
\title{FUNCTION: Numerical prediction graph}
\usage{
graph_npreds(data)
}
\description{
This function takes in the results of
running predictions and then plots
the comparison of the actual mutation score
to the predicted mutation score
}

