% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_overall}
\alias{graph_overall}
\title{FUNCTION: graph_overall}
\usage{
graph_overall(data, ytitle, title, x_string = "Algorithm",
  y_string = "value")
}
\description{
This graphing function creates a single box and whisker plot to visualize the
results from running the benchmark function in mspredictr
}

