% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_preds}
\alias{graph_preds}
\title{FUNCTION: graph_preds}
\usage{
graph_preds(data, ytitle, title, x_string = "Algorithm", y_string = "value")
}
\description{
This function creates a faceted box and whisker plot, based on the program, from the results returned by the
benchmark function in mspredictr
}

