% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{npred_by_program}
\alias{npred_by_program}
\title{FUNCTION: npred_by_program}
\usage{
npred_by_program(method = "rf", pred = "bpmail", data = totalCDNoNEvo)
}
\arguments{
\item{method}{Machine learning algorithm}

\item{pred}{Program to predict mutation scores}

\item{data}{Dataframe with metrics and mutation scores}
}
\description{
Numerical prediction of Mutation Scores
based on source code metrics
}

