% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-pred.R
\name{num_predict}
\alias{num_predict}
\title{FUNCTION: num_predict}
\usage{
num_predict(data, method, pred)
}
\description{
This function takes in data and
predicts the desired value using
the desired machine learning algorithm
}

