% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{pred_bench}
\alias{pred_bench}
\title{FUNCTION: pred_bench}
\usage{
pred_bench(trials, pred, data, methods, sampling, type, kfolds, repeats)
}
\description{
This function runs multiple trials of experimentation on a list of machine learning algorithm
using the same data
}

