% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{pred_frame}
\alias{pred_frame}
\title{FUNCTION: pred_frame}
\usage{
pred_frame(method, pred, data, sampling, type, kfolds, repeats, eval)
}
\description{
This function provides a frame for all possible prediction
configurations
}

