% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{benchmark}
\alias{benchmark}
\title{FUNCTION: benchmark}
\usage{
benchmark(trials, methods, programs, data = totalCDCatEvo,
  sampling = "regular", class = NULL)
}
\description{
This function runs multiple trials of experimentation
on a list of programs using a list of
different machine learning algorithms
}

